# wrapped 32-bit products made four decisions come out backwards on large inputs
Each of these multiplies a length by a small constant in int, and the wrapped
negative product silently reverses the test it feeds. None of them throws.

BPE stopped merging after one round. The progress test (prevLen - dataLen) * 200
wraps once a single round removes more than 10,737,418 pairs, which the top bigram
of a large text easily reaches, and the wrapped value reads as no progress. Costs
compression ratio with no other symptom.

ZIP Reduce returned an all-zero buffer of exactly the right length. Its bit total
compressed.Length * 8 wraps at 2^28 compressed bytes, so the decode loop never ran.
Its intermediate-list capacity originalSize * 2 wraps at 2^30 as well.

MS-LZH never emitted a byte. Its decode budget originalSize * 8 + 1024 goes negative
at 2^28, so the symbol loop exited immediately and the outer loop spun until the bit
reader hit the end of the stream.

Brotli mis-padded its uncompressed meta-blocks. BitLength() counts the whole output
stream, so past 2^28 bytes of output the byte-alignment padding computed from it was
wrong.

In Compression.Analysis, SuccessEvaluator rejected every candidate as excessive
expansion above 20.5 MB, LzwHeuristic and TrialHuffman reported no match above
256 MB, and TrialDecompressor and ChainReconstructor disabled every trial strategy
above 512 MB.

All taken in 64-bit. Below each wrap point the widened arithmetic is exactly the old
arithmetic, so output is unchanged.

+++ Compression.Analysis/ChainReconstruction/ChainReconstructor.cs
@@ -113,7 +113,7 @@ public sealed class ChainReconstructor {
 
       try {
         using var cts = new CancellationTokenSource(BlindTrialTimeoutMs(current.Length));
-        var attempt = strategy.TryDecompress(current, Math.Min(current.Length * 4, 1024 * 1024), cts.Token);
+        var attempt = strategy.TryDecompress(current, (int)Math.Min((long)current.Length * 4, 1024 * 1024), cts.Token);
         if (!attempt.Success || attempt.Output == null) continue;
 
         var quality = SuccessEvaluator.Evaluate(current, attempt.Output, strategy.Algorithm);
